video.report method
